% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMP}
\alias{plotMP}
\title{plotMP}
\usage{
plotMP(d, ytitle = "Age cal. BP")
}
\arguments{
\item{d}{the output of MarcoPolo application}

\item{ytitle}{y-axis title, defaults to "Age cal. BP"}
}
\value{
A plot object
}
\description{
Plots landcover reconstructed with MarcoPolo
}
\examples{
## Apply MarcoPolo on package data (from Mrotzek et al. 2017) and plot:
data("MP_example")

## run MarcoPolo
a <- MarcoPolo(localMP, regionalMP, paramsMP, assignmentMP)

## prepare plot
p <- plotMP(a, ytitle = "Site")

# and write to pdf file
pdf("MarcoPolo plot.pdf", height = 12, width = 12)
p
dev.off()

}
\references{
Mrotzek A., Couwenberg J., Theuerkauf M. and Joosten H. (2017) MARCO POLO –
  A new and simple tool for pollen-based stand-scale vegetation reconstruction. The Holocene 27: 321-330.
  DOI:10.1177/0959683616660171
}
\author{
Martin Theuerkauf <martin.theuerkauf@uni-greifswald.de>

Almut Mrotzek <almut.mrotzek@gmx.de>
}
