% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REVEALS.R
\name{REVEALSinR}
\alias{REVEALSinR}
\title{REVEALSinR}
\usage{
REVEALSinR(pollen, params, tBasin, dBasin, dwm = "lsm unstable", n = 1000,
  regionCutoff = 1e+05, ppefun = rnorm_reveals,
  pollenfun = rmultinom_reveals, writeresults = FALSE, verbose = TRUE)
}
\arguments{
\item{pollen}{data frame with raw pollen counts. Counts are arranged as
pollen taxa in columns and samples in rows. The first colum includes sample
ages or depth, i.e. only numeric values! The first row includes taxon
names. Taxon names cannot include spaces! See 'pollenTS' in the 'Tiefer
See' data set to as example. For each pollen taxon, fallspeed and PPEs have
to be provided in params}

\item{params}{data frame of parameters. Includes 'fallspeed' of pollen (in m
s-1), relative pollen productivity estimates ('PPE') and their standard
error. Pollen taxa are given in rows, parameters in colums.  Column names
must be "species", "fallspeed", "PPE", and "PPE.error". See 'paramsTS' in
the 'Tiefer See' data set as example. For each pollen taxon in the pollen
data, one record of parameters with exactly the same name is required.
Params may include more taxa, also in different order, so that a standard
list may be established.}

\item{tBasin}{type of basin, either 'lake' oder 'peatland'}

\item{dBasin}{basin diameter in meter}

\item{dwm}{distance weighting method. The following methods are implemented:
'lsm unstable', 'gpm unstable', 'gpm neutral'. 'lsm' refers to the the
Lagrangian stochastic model presented by Kuparinen et al. 2007. 'gpm'
refers to the Gaussian plume model, based on Sutton's equations, used by
Prentice (1985; Quaternary Research 23: 76-86). Further methods can be
added.}

\item{n}{number of model runs per time slice, by default 1000}

\item{regionCutoff}{diameter of the reconstruction region in meters, by
default 100000}

\item{ppefun}{function to randomise these parameters, defaults to
rnorm_reveals}

\item{pollenfun}{function to randomise pollen counts, for instance rMaher or
rmultinom_reveals}

\item{writeresults}{boolean, indicates whether to write results to csv-files
or to return them on the command-line}

\item{verbose}{boolean, indicating whether to print progress messages}
}
\value{
REVEALS returns a list with details of the reconstructed vegetation
  composition, calculated from the n model runs for each sample

...meansim:	mean

...mediansim: median

...q90sim:	90\% quantile

...q10sim: 10\% quantile

...sdsim: standard deviation
}
\description{
The REVEALS model (Regional Estimates of VEgetation Abundance from Large
Sites), introduced by Shinya Sugita in 2007, aims to translate pollen data
from large sites into regional vegetation composition. The model applies
relative pollen productivity estimates (known as PPEs or RPPs) to account for
the productivity bias and a "dispersal and deposition factor" K to account
for the dispersal bias in pollen data. K is calculated with a dispersal
model. REVEALS does not account for the homogeneity bias; we thus have to
assume that the vegetation of the pollen source area has been homogeneous.
The present implementation of REVEALS in R adds full flexibility with respect
to the underlying dispersal model. REVEALSinR by default uses a Lagrangian
stochastic (LS) dispersal model (Kuparinen et al. 2007) because this model
predicts the pollen dispersal pattern more realistically than the simple
Gaussian models used hitherto. Model selection has large effects on the
results. REVEALSinR provides further options for testing. To arrive at error
estimates, model runs are repeated (by default 1000 times) with random error
added in pollen data and PPEs during each model run. To account for pollen
deposition in lakes, a lake model is included.
}
\details{
REVEALSinR applies calculations on a sequence of samples, such as a pollen
record. When needed, the REVEALS function takes calculations for a single
sample.
}
\examples{
## Application on a core section from Lake Tiefer See,
## covering the period 1880 t0 2010 CE:
data("Tiefer_See", package = "disqover")

## inspect the sample data
pollenTS
paramsTS

## and apply REVEALS...
cover <- REVEALSinR(pollenTS, paramsTS, dwm='lsm unstable', tBasin='lake', dBasin=600)

## plot results as stratigraphic diagram ('pollen diagram')
plotREVEALS(cover, type = "single", ytitle = "Age CE")

## plot results as stacked area diagram
plotREVEALS(cover, type = "stacked", ytitle = "Age CE")

## load pollen data from an Excel file (e.g.'Pollen_Tiefer_See_1880-2010.xlsx')
datafile <- system.file('extdata/reveals', 'Pollen_Tiefer_See_1880-2010.xlsx', package = 'disqover') # with data file from the package
datafile <- 'YourFileFolder/Pollen_Tiefer_See_1880-2010.xlsx' # with the file in YourFileFolder
pollenTS <- as.data.frame(do.call(read_excel,args=c(list(datafile,sheet = 1)))) # function reads first worksheet

## load parameters from an Excel file (e.g.'pf_PPE_MV2015.xlsx')
datafile <- system.file('extdata/reveals', 'pf_PPE_MV2015.xlsx', package = 'disqover') # with data file from the package
datafile <- 'YourFileFolder/pf_PPE_MV2015.xlsx' # with the file in YourFileFolder
paramsTS <- as.data.frame(do.call(read_excel,args=c(list(datafile,sheet = 1)))) #function reads first worksheet

}
\references{
Theuerkauf M., Couwenberg J., Kuparinen A., Liebschar V. (2016) A
  matter of dispersal: REVEALSinR introduces state-of-the-art dispersal
  models to quantitative vegetation reconstruction. Vegetation History and
  Archaeobotany 25: 541-553. https://doi.org/10.1007/s00334-016-0572-0
}
\author{
Martin Theuerkauf <martin.theuerkauf@uni-greifswald.de>

Volkmar Liebscher <volkmar.liebscher@uni-greifswald.de>
}
